Performed some refactoring
