+++ lib/features/shop/screens/home/home.dart
@@ -1,12 +1,9 @@
 import 'package:e_commerce_app/features/shop/screens/home/widgets/home_appbar.dart';
-import 'package:e_commerce_app/utils/constants/colors.dart';
 import 'package:e_commerce_app/utils/constants/sizes.dart';
-import 'package:e_commerce_app/utils/device/device_utility.dart';
-import 'package:e_commerce_app/utils/helpers/helper_functions.dart';
 import 'package:flutter/material.dart';
-import 'package:iconsax/iconsax.dart';
 
 import '../../../../common/widgets/primary_header_container.dart';
+import '../../../../common/widgets/search_container.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -27,61 +24,13 @@ class HomeScreen extends StatelessWidget {
                   const SizedBox(height: AppSizes.spaceBetweenSections),
 
                   //Searchbar
-                  AppSearchContainer(
-                    text: 'Search in store',
-                    icon: Iconsax.search_normal,
-                    showBackground: true,
-                    showBorder: true,
-                  ),
+                  AppSearchContainer(text: 'Search in store'),
+                  const SizedBox(height: AppSizes.spaceBetweenSections),
 
                   //Categories
                 ],
               ),
             ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class AppSearchContainer extends StatelessWidget {
-  const AppSearchContainer({
-    super.key,
-    required this.text,
-    this.icon,
-    required this.showBackground,
-    required this.showBorder,
-  });
-
-  final String text;
-  final IconData? icon;
-  final bool showBackground, showBorder;
-
-  @override
-  Widget build(BuildContext context) {
-    final dark = AppHelperFunctions.isDarkMode(context);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: AppSizes.defaultSpace),
-      child: Container(
-        width: AppDeviceUtils.getScreenWidth(context),
-        padding: const EdgeInsets.all(AppSizes.md),
-        decoration: BoxDecoration(
-          color:
-              showBackground
-                  ? (dark ? AppColors.dark : AppColors.light)
-                  : Colors.transparent,
-          borderRadius: BorderRadius.circular(AppSizes.cardRadiusLg),
-          border: showBorder ? Border.all(color: AppColors.grey) : null,
-        ),
-        child: Row(
-          children: [
-            Icon(icon, color: AppColors.darkerGrey),
-
-            const SizedBox(width: AppSizes.spaceBetweenItems),
-
-            Text(text, style: Theme.of(context).textTheme.bodySmall),
           ],
         ),
       ),
